Add Spread Pattern Attribute to WeaponProjectileData and add Random Spread Pattern

+++ src/game/resources.rs
@@ -12,7 +12,7 @@ pub struct GameParametersResource {
     pub camera_zoom_out_scale: f32,
     /// Maximum possible projectiles for 1 of the player/mobs shots. Mainly kept low for perf and as
     /// a hard cap (along with fire rate) on how much of a "bullet hell" each mob/player creates.
-    pub max_player_projectiles: f32,
+    pub max_player_projectiles: u16,
     /// Maximum possible speed of an entity
     pub max_speed: f32,
     /// Maximum angle between the first and last projectile
